style, move createpost
